fix: bug in task Id 25 "user data is refetched every time" at ProfilePage
1. fix bug
2. implemented Skeleton at TelegramProfile
3. fix bug at userCurrencies "don`t hide settings button under currencies after add new currencies"
