+++ src/components/UserProfile/Skeleton/CurrenciesSkeleton.js
@@ -1,12 +1,5 @@
-import {
-  Tooltip,
-  Chip,
-  Stack,
-  Grid,
-  Divider,
-  Skeleton,
-  Button,
-} from '@mui/material';
+import { Tooltip, Chip, Stack, Grid } from '@mui/material';
+import renderGridContent from './renderGridContent';
 
 function CurrenciesSkeleton() {
   const renderTooltips = () => {
@@ -24,27 +17,7 @@ function CurrenciesSkeleton() {
     }
     return tooltips;
   };
-  const renderGridContent = () => {
-    const grid = [];
-    for (let i = 0; i < 2; i++) {
-      grid.push(
-        <Grid item xs={12}>
-          <Stack
-            direction='row'
-            spacing={2}
-            sx={{ alignItems: 'center' }}
-            divider={<Divider orientation='vertical' flexItem />}
-          >
-            <Skeleton variant='text' sx={{ width: '200px' }} />
-            <Button>
-              <Skeleton variant='rounded' sx={{ width: '100%' }} />
-            </Button>
-          </Stack>
-        </Grid>,
-      );
-    }
-    return grid;
-  };
+
   return (
     <>
       <Grid
@@ -74,7 +47,7 @@ function CurrenciesSkeleton() {
         direction='column'
         sx={{ textAlign: 'left', marginTop: '5px' }}
       >
-        {renderGridContent()}
+        {renderGridContent(2)}
       </Grid>
     </>
   );
