+++ src/components/UserProfile/TelegramProfile.js
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useEffect, useState, useMemo } from 'react';
 import {
   TextField,
   FormControl,
@@ -22,8 +22,9 @@ import { validationProps } from '../../utils/validationForm';
 import { useUpdateTgUserName, useUpdateUserSettings } from '../../queries';
 import TransitionsModal from '../modal/modal';
 import { showSnackbarMessage } from '../../store/slices/snackbarMessage';
+import TelegramSkeleton from './Skeleton/TelegramSkeleton';
 
-function TelegramProfile({ enableTgNotifications, telegramUser }) {
+function TelegramProfile({ currentUser }) {
   const { mutateAsync: putTgUserName, isLoading: loadingTgUsername } =
     useUpdateTgUserName();
   const { mutateAsync: putUserSettings, isLoading: loadingUserSettings } =
@@ -35,7 +36,11 @@ function TelegramProfile({ enableTgNotifications, telegramUser }) {
   const [isNotificationOn, setIsNotificationOn] = useState(true);
   const [tgCollapseOff, setTgCollapseOff] = useState(false);
   const dispatch = useDispatch();
-
+  const showSkeleton = useMemo(() => !currentUser, [currentUser]);
+  const { enableTgNotifications, telegramUser } = useMemo(() => {
+    if (currentUser) return currentUser;
+    return [];
+  }, [currentUser]);
   const updateTg = async (Name) => {
     try {
       await putTgUserName(Name);
@@ -112,168 +117,172 @@ function TelegramProfile({ enableTgNotifications, telegramUser }) {
 
   return (
     <form onSubmit={updateTgUserName}>
-      <Grid
-        container
-        spacing={2}
-        direction='column'
-        sx={{ textAlign: 'left', marginTop: '5px' }}
-      >
-        <TransitionsModal
-          isOpen={confirmModalOpen}
-          title='Выключить уведомления?'
-          handleClose={() => {
-            setConfirmModalOpen(false);
-          }}
-          body={
-            <Stack
-              direction='row'
-              spacing={2}
-              sx={{ alignItems: 'center', marginTop: '10px' }}
-            >
-              <LoadingButton
-                loading={loadingUserSettings}
-                variant='contained'
-                color='success'
-                type='submit'
-                endIcon={<CheckIcon />}
-                onClick={() => updateUserSettings()}
-              >
-                Да
-              </LoadingButton>
-              <Button
-                variant='outlined'
-                color='error'
-                onClick={() => {
-                  setConfirmModalOpen(false);
-                }}
-                endIcon={<ClearIcon />}
-              >
-                Нет
-              </Button>
-            </Stack>
-          }
-        />
-        {telegramUser && (
-          <>
-            <Grid item xs={12}>
-              <Typography
-                variant='body1'
-                sx={{ width: '200px', fontWeight: 'bold' }}
-              >
-                {`@${telegramUser}`}
-              </Typography>
-            </Grid>
-            <Grid item xs={12}>
+      {showSkeleton ? (
+        <TelegramSkeleton />
+      ) : (
+        <Grid
+          container
+          spacing={2}
+          direction='column'
+          sx={{ textAlign: 'left', marginTop: '5px' }}
+        >
+          <TransitionsModal
+            isOpen={confirmModalOpen}
+            title='Выключить уведомления?'
+            handleClose={() => {
+              setConfirmModalOpen(false);
+            }}
+            body={
               <Stack
                 direction='row'
                 spacing={2}
-                sx={{ alignItems: 'center' }}
-                divider={<Divider orientation='vertical' flexItem />}
+                sx={{ alignItems: 'center', marginTop: '10px' }}
               >
-                <Typography variant='body1' sx={{ width: '200px' }}>
-                  Уведомления
-                </Typography>
-                <Switch
-                  checked={isNotificationOn}
-                  onChange={toggleNotifications}
-                  inputProps={{ 'aria-label': 'controlled' }}
-                />
+                <LoadingButton
+                  loading={loadingUserSettings}
+                  variant='contained'
+                  color='success'
+                  type='submit'
+                  endIcon={<CheckIcon />}
+                  onClick={() => updateUserSettings()}
+                >
+                  Да
+                </LoadingButton>
+                <Button
+                  variant='outlined'
+                  color='error'
+                  onClick={() => {
+                    setConfirmModalOpen(false);
+                  }}
+                  endIcon={<ClearIcon />}
+                >
+                  Нет
+                </Button>
               </Stack>
-            </Grid>
-          </>
-        )}
-        <Grid item xs={12}>
-          <Stack
-            direction='row'
-            spacing={2}
-            sx={{ alignItems: 'center' }}
-            divider={<Divider orientation='vertical' flexItem />}
-          >
-            <Typography variant='body1' sx={{ width: '200px' }}>
-              Привязать {telegramUser && 'другой'} аккаунт
-            </Typography>
-            <Button
-              onClick={() => {
-                setTgCollapseOff(!tgCollapseOff);
-              }}
-            >
-              <TelegramIcon />
-            </Button>
-          </Stack>
-        </Grid>
-        <Grid item xs={12}>
-          <Collapse in={tgCollapseOff}>
-            <Grid
-              container
-              spacing={2}
-              direction='column'
-              sx={{ textAlign: 'left', marginTop: '5px' }}
-            >
+            }
+          />
+          {telegramUser && (
+            <>
               <Grid item xs={12}>
                 <Typography
                   variant='body1'
-                  gutterBottom
-                  sx={{ textAlign: 'left' }}
+                  sx={{ width: '200px', fontWeight: 'bold' }}
                 >
-                  Для привязки аккаунта, введите Имя пользователя tg и напишите
-                  любое сообщение нашему чат боту.
+                  {`@${telegramUser}`}
                 </Typography>
-              </Grid>
-              <Grid item xs={12}>
-                <FormControl fullWidth required>
-                  <TextField
-                    variant='outlined'
-                    label='Имя пользователя'
-                    value={enteredTgName}
-                    type='text'
-                    onChange={(e) => {
-                      setIsTgError(false);
-                      setEnteredTgName(e.target.value);
-                    }}
-                    id='enteredTgName'
-                    required
-                    helperText={choseTgNameTextHelper()}
-                    className={enteredTgName && classes.valid}
-                    error={isTgError}
-                    onFocus={() => setEnteredTgName('@')}
-                    inputProps={{
-                      inputMode: 'text',
-                      pattern: tgName.inputPropsPattern,
-                      title: tgName.errorPatternTitle,
-                    }}
-                  />
-                </FormControl>
               </Grid>
               <Grid item xs={12}>
                 <Stack
                   direction='row'
                   spacing={2}
                   sx={{ alignItems: 'center' }}
+                  divider={<Divider orientation='vertical' flexItem />}
                 >
-                  <Button
-                    variant='outlined'
-                    onClick={() => {
-                      window.open('https://t.me/bablo_project_bot', '_blank');
-                    }}
-                    endIcon={<TelegramIcon />}
-                  >
-                    Telegram - bot
-                  </Button>
-                  <LoadingButton
-                    loading={loadingTgUsername}
-                    variant='contained'
-                    color='success'
-                    type='submit'
-                    endIcon={<CheckIcon />}
-                  >
-                    Сохранить
-                  </LoadingButton>
+                  <Typography variant='body1' sx={{ width: '200px' }}>
+                    Уведомления
+                  </Typography>
+                  <Switch
+                    checked={isNotificationOn}
+                    onChange={toggleNotifications}
+                    inputProps={{ 'aria-label': 'controlled' }}
+                  />
                 </Stack>
               </Grid>
-            </Grid>
-          </Collapse>
+            </>
+          )}
+          <Grid item xs={12}>
+            <Stack
+              direction='row'
+              spacing={2}
+              sx={{ alignItems: 'center' }}
+              divider={<Divider orientation='vertical' flexItem />}
+            >
+              <Typography variant='body1' sx={{ width: '200px' }}>
+                Привязать {telegramUser && 'другой'} аккаунт
+              </Typography>
+              <Button
+                onClick={() => {
+                  setTgCollapseOff(!tgCollapseOff);
+                }}
+              >
+                <TelegramIcon />
+              </Button>
+            </Stack>
+          </Grid>
+          <Grid item xs={12}>
+            <Collapse in={tgCollapseOff}>
+              <Grid
+                container
+                spacing={2}
+                direction='column'
+                sx={{ textAlign: 'left', marginTop: '5px' }}
+              >
+                <Grid item xs={12}>
+                  <Typography
+                    variant='body1'
+                    gutterBottom
+                    sx={{ textAlign: 'left' }}
+                  >
+                    Для привязки аккаунта, введите Имя пользователя tg и
+                    напишите любое сообщение нашему чат боту.
+                  </Typography>
+                </Grid>
+                <Grid item xs={12}>
+                  <FormControl fullWidth required>
+                    <TextField
+                      variant='outlined'
+                      label='Имя пользователя'
+                      value={enteredTgName}
+                      type='text'
+                      onChange={(e) => {
+                        setIsTgError(false);
+                        setEnteredTgName(e.target.value);
+                      }}
+                      id='enteredTgName'
+                      required
+                      helperText={choseTgNameTextHelper()}
+                      className={enteredTgName && classes.valid}
+                      error={isTgError}
+                      onFocus={() => setEnteredTgName('@')}
+                      inputProps={{
+                        inputMode: 'text',
+                        pattern: tgName.inputPropsPattern,
+                        title: tgName.errorPatternTitle,
+                      }}
+                    />
+                  </FormControl>
+                </Grid>
+                <Grid item xs={12}>
+                  <Stack
+                    direction='row'
+                    spacing={2}
+                    sx={{ alignItems: 'center' }}
+                  >
+                    <Button
+                      variant='outlined'
+                      onClick={() => {
+                        window.open('https://t.me/bablo_project_bot', '_blank');
+                      }}
+                      endIcon={<TelegramIcon />}
+                    >
+                      Telegram - bot
+                    </Button>
+                    <LoadingButton
+                      loading={loadingTgUsername}
+                      variant='contained'
+                      color='success'
+                      type='submit'
+                      endIcon={<CheckIcon />}
+                    >
+                      Сохранить
+                    </LoadingButton>
+                  </Stack>
+                </Grid>
+              </Grid>
+            </Collapse>
+          </Grid>
         </Grid>
-      </Grid>
+      )}
     </form>
   );
 }
